pyproject.toml
[tool.poetry]
name = "deep-search-lite"
version = "0.2.0"
description = "Introducing a streamlined and efficient library designed for Similarity Search, crafted to seamlessly integrate with any feature encoder of your choice. This library empowers you to encode your data effortlessly, regardless of its nature, facilitating the retrieval of closely matched objects with precision. Developed with a keen focus on deployment efficiency, our library boasts a lightweight architecture and relies on a minimal set of dependencies, ensuring a lean and robust solution for your similarity search needs. Whether you're looking to enhance data analysis, improve content discovery, or power recommendation systems, this library provides the foundational tools you need to achieve fast, accurate results."
authors = ["Ibad Rather <ibad.rather.ir@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
faiss-cpu = "^1.7.4"
concurrent-log-handler = "^0.9.25"
pathlib = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest = "^8.0.0"
pre-commit = "^3.6.1"
reorder-python-imports = "^3.12.0"
pillow = "^10.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''
